Alterações pós-reunião com a Proteção Civil

Fixes diversos
Switch de desbloqueio de stock
Modal regulamento distribuições
Permissões frontend
Fita do Tempo
Resolução de problemas com datas (datetime para timestampz)

+++ frontend/app/components/incidentPriorities/DeleteModal.vue
@@ -1,5 +1,6 @@
 <script setup lang="ts">
 import { useApiStore } from "@/stores/api"
+import {useAuthStore} from "~/stores/auth";
 
 const api = useApiStore()
 const toast = useToast()
@@ -18,7 +19,18 @@ const openModel = computed({
 })
 
 const onSubmit = async () => {
+  if (!useAuthStore().hasPermission('INCIDENT_PRIORITIES_DELETE')) return;
+
   if (!props.id) return
+
+  if (!await checkServerAccess()) {
+    useToast().add({
+      title: 'Sem ligação à internet!',
+      description: 'Não é possível guardar alterações sem estar ligado à internet. Tente novamente mais tarde',
+      color: 'error'
+    });
+    return;
+  }
 
   try {
     await api.deleteIncidentPriority(props.id)
